student-dropout-app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.2
pandas>=2.2.2
scikit-learn>=1.4.2
joblib>=1.4.2
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.1
